refactor: seperate local blockchain setup and helper methods from smart contracts
