+++ examples/sudoku/ts/src/sudoku-zkapp.ts
@@ -5,22 +5,17 @@ import {
   DeployArgs,
   SmartContract,
   method,
-  PrivateKey,
-  Mina,
   Bool,
   state,
   State,
   isReady,
   Poseidon,
-  Party,
   Permissions,
 } from 'snarkyjs';
 
-export { deploy, submitSolution, getZkAppState };
-
 await isReady;
 
-class Sudoku extends CircuitValue {
+export class Sudoku extends CircuitValue {
   @matrixProp(Field, 9, 9) value: Field[][];
 
   constructor(value: number[][]) {
@@ -103,60 +98,8 @@ export class SudokuZkApp extends SmartContract {
   }
 }
 
-// setup
-const Local = Mina.LocalBlockchain();
-Mina.setActiveInstance(Local);
-const account1 = Local.testAccounts[0].privateKey;
-
-const zkAppPrivateKey = PrivateKey.random();
-let zkAppAddress = zkAppPrivateKey.toPublicKey();
-
-async function deploy(sudoku: number[][]) {
-  let tx = await Mina.transaction(account1, () => {
-    Party.fundNewAccount(account1);
-    let zkApp = new SudokuZkApp(zkAppAddress);
-    let sudokuInstance = new Sudoku(sudoku);
-    zkApp.deploy({ zkappKey: zkAppPrivateKey });
-    zkApp.setPermissions({
-      ...Permissions.default(),
-      editState: Permissions.proofOrSignature(),
-    });
-    zkApp.sudokuHash.set(sudokuInstance.hash());
-    zkApp.isSolved.set(Bool(false));
-  });
-  await tx.send().wait();
-}
-
-async function submitSolution(sudoku: number[][], solution: number[][]) {
-  let tx = await Mina.transaction(account1, () => {
-    let zkApp = new SudokuZkApp(zkAppAddress);
-    zkApp.submitSolution(new Sudoku(sudoku), new Sudoku(solution));
-    zkApp.self.sign(zkAppPrivateKey);
-    zkApp.self.body.incrementNonce = Bool(true);
-  });
-  try {
-    await tx.send().wait();
-    return true;
-  } catch (err) {
-    return false;
-  }
-}
-
-async function getZkAppState() {
-  let zkAppState = Mina.getAccount(zkAppAddress).zkapp?.appState;
-  if (zkAppState === undefined)
-    throw Error('Account does not have zkApp state.');
-  let sudokuHash = fieldToHex(zkAppState?.[0]);
-  let isSolved = zkAppState[1].equals(true).toBoolean();
-  return { sudokuHash, isSolved };
-}
-
 // helpers
 function divmod(k: number, n: number) {
   let q = Math.floor(k / n);
   return [q, k - q * n];
-}
-
-function fieldToHex(field: Field) {
-  return BigInt(field.toString()).toString(16);
 }
